backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
fastapi==0.104.1
h11==0.16.0
idna==3.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.8.0
python-dotenv==1.1.0
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
